// backend/src/schemas/auth.schema.ts
import { z } from 'zod'

export const registerSchema = z.object({
  body: z.object({
    email: z.email(),
    password: z.string().min(8),
    fullName: z.string().min(2).max(100).trim(),
  }),
})

export const loginSchema = z.object({
  body: z.object({
    email: z.email(),
    password: z.string().min(1),
  }),
})

export type RegisterInput = z.infer<typeof registerSchema>['body']
export type LoginInput = z.infer<typeof loginSchema>['body']